final refactor and test updates
