+++ ppr-ui/src/components/parties/party/index.ts
@@ -2,4 +2,5 @@ export { default as EditParty } from './EditParty.vue'
 export { default as PartyAutocomplete } from './PartyAutocomplete.vue'
 export { default as PartySearch } from './PartySearch.vue'
 export { default as RegisteringParty } from './RegisteringParty.vue'
+export { default as RegisteringPartyChange } from './RegisteringPartyChange.vue'
 export { default as SecuredParties } from './SecuredParties.vue'
